Refactorizar la estructura del paquete: consolidar interfaces y mejorar la organización de los modelos de autenticación y chat
