+++ composeApp/src/androidMain/kotlin/org/connexuss/project/http/androidHttpClientFactory.kt
@@ -1,12 +1,12 @@
-package org.connexuss.project
+package org.connexuss.project.http
 /*
-import io.ktor.client.*
-import io.ktor.client.engine.cio.*
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.okhttp.OkHttp
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.serialization.kotlinx.json.json
 import kotlinx.serialization.json.Json
 
-actual fun createHttpClient(): HttpClient = HttpClient(CIO) {
+actual fun createHttpClient(): HttpClient = HttpClient(OkHttp) {
     install(ContentNegotiation) {
         json(Json {
             ignoreUnknownKeys = true
