--- composeApp/src/commonMain/kotlin/org/connexuss/project/acceso/RecoveryRequest.kt
@@ -1,5 +1,0 @@
-package org.connexuss.project.acceso
-
-class RecoveryRequest {
-    val email: String = ""
-}
