--- composeApp/src/commonMain/kotlin/org/connexuss/project/acceso/RecoveryResult.kt
@@ -1,6 +1,0 @@
-package org.connexuss.project.acceso
-
-class RecoveryResult {
-    val exito: Boolean = false
-    val mensaje: String? = null
-}
