--- composeApp/src/commonMain/kotlin/org/connexuss/project/comunicacion/ChatRepository.kt
@@ -1,7 +1,0 @@
-package org.connexuss.project.comunicacion
-
-interface ChatRepository {
-    suspend fun sendMessage(roomId: String, message: Mensaje): Boolean
-    suspend fun getChatRoom(roomId: String): Conversacion?
-    // Podrías agregar funciones para crear salas, listar salas, etc.
-}
