--- composeApp/src/commonMain/kotlin/org/connexuss/project/encriptacion/UtilidadesEncriptacion.kt
@@ -1,6 +1,0 @@
-package org.connexuss.project.encriptacion
-
-// funciones de encriptacion hash (Resumen)
-fun hash(texto: String): String {
-    return texto.hashCode().toString()
-}
