+++ composeApp/src/commonMain/kotlin/org/connexuss/project/firebase/Firebase.kt
@@ -1,0 +1,83 @@
+package org.connexuss.project.firebase
+
+import org.connexuss.project.comunicacion.PeticionRegistro
+import org.connexuss.project.comunicacion.PeticionLogin
+import org.connexuss.project.usuario.Usuario
+
+object FirebaseAuthManager {
+    suspend fun registerUser(request: PeticionRegistro): Boolean {
+        // Lógica para registrar usuario usando Firebase Authentication
+        // Por ejemplo, llamar a FirebaseAuth.createUserWithEmailAndPassword(...)
+        return true // Simulación
+    }
+
+    suspend fun loginUser(request: PeticionLogin): Boolean {
+        // Lógica para iniciar sesión usando Firebase Authentication
+        return true // Simulación
+    }
+
+    suspend fun recoverPassword(email: String): Boolean {
+        // Lógica para enviar correo de recuperación usando Firebase Authentication
+        return true // Simulación
+    }
+}
+
+object FirebaseConfig {
+    fun initFirebase() {
+        // Inicialización de Firebase (en Android se llama FirebaseApp.initializeApp(context))
+        // En iOS o web, usar los métodos correspondientes
+    }
+}
+
+object FirebaseConstants {
+    const val COLLECTION_USUARIOS = "usuarios"
+    const val COLLECTION_CHAT = "chat"
+    const val COLLECTION_FORO = "foro"
+    // Otras constantes necesarias...
+}
+
+object FirebaseFirestoreManager {
+    suspend fun createUser(usuario: Usuario): Boolean {
+        // Lógica para almacenar el usuario en Firestore usando FirebaseConstants.COLLECTION_USUARIOS
+        return true // Simulación
+    }
+
+    suspend fun getUserByEmail(email: String): Usuario? {
+        // Lógica para buscar un usuario por correo en Firestore
+        return null // Simulación
+    }
+
+    suspend fun updateUserPassword(userId: String, newPassword: String): Boolean {
+        // Lógica para actualizar el campo de contraseña (o alias privado) en Firestore
+        return true // Simulación
+    }
+
+    // Otros métodos para chats, hilos de foro, etc.
+}
+
+object FirebaseFunctionsManager {
+    suspend fun callFunction(name: String, data: Map<String, Any>): Any? {
+        // Lógica para invocar una función en Firebase Functions vía HTTP o SDK
+        return null // Simulación
+    }
+}
+
+object FirebaseRemoteConfigHelper {
+    suspend fun fetchConfig(): Map<String, Any> {
+        // Lógica para obtener valores de Remote Config
+        return emptyMap()
+    }
+}
+
+object FirebaseStorageManager {
+    suspend fun uploadFile(path: String, data: ByteArray): String {
+        // Lógica para subir un archivo a Firebase Storage
+        // Devuelve la URL del archivo
+        return "url_del_archivo"
+    }
+
+    suspend fun downloadFile(path: String): ByteArray? {
+        // Lógica para descargar un archivo de Firebase Storage
+        return null
+    }
+}
