--- composeApp/src/commonMain/kotlin/org/connexuss/project/firebase/FirebaseAuthManager.kt
@@ -1,22 +1,0 @@
-package org.connexuss.project.firebase
-
-import org.connexuss.project.acceso.RegistroRequest
-import org.connexuss.project.acceso.LoginRequest
-
-object FirebaseAuthManager {
-    suspend fun registerUser(request: RegistroRequest): Boolean {
-        // Lógica para registrar usuario usando Firebase Authentication
-        // Por ejemplo, llamar a FirebaseAuth.createUserWithEmailAndPassword(...)
-        return true // Simulación
-    }
-
-    suspend fun loginUser(request: LoginRequest): Boolean {
-        // Lógica para iniciar sesión usando Firebase Authentication
-        return true // Simulación
-    }
-
-    suspend fun recoverPassword(email: String): Boolean {
-        // Lógica para enviar correo de recuperación usando Firebase Authentication
-        return true // Simulación
-    }
-}
