--- composeApp/src/commonMain/kotlin/org/connexuss/project/firebase/FirebaseConfig.kt
@@ -1,8 +1,0 @@
-package org.connexuss.project.firebase
-
-object FirebaseConfig {
-    fun initFirebase() {
-        // Inicialización de Firebase (en Android se llama FirebaseApp.initializeApp(context))
-        // En iOS o web, usar los métodos correspondientes
-    }
-}
