--- composeApp/src/commonMain/kotlin/org/connexuss/project/firebase/FirebaseConstants.kt
@@ -1,8 +1,0 @@
-package org.connexuss.project.firebase
-
-object FirebaseConstants {
-    const val COLLECTION_USUARIOS = "usuarios"
-    const val COLLECTION_CHAT = "chat"
-    const val COLLECTION_FORO = "foro"
-    // Otras constantes necesarias...
-}
