--- composeApp/src/commonMain/kotlin/org/connexuss/project/firebase/FirebaseFirestoreManager.kt
@@ -1,22 +1,0 @@
-package org.connexuss.project.firebase
-
-import org.connexuss.project.usuario.Usuario
-
-object FirebaseFirestoreManager {
-    suspend fun createUser(usuario: Usuario): Boolean {
-        // Lógica para almacenar el usuario en Firestore usando FirebaseConstants.COLLECTION_USUARIOS
-        return true // Simulación
-    }
-
-    suspend fun getUserByEmail(email: String): Usuario? {
-        // Lógica para buscar un usuario por correo en Firestore
-        return null // Simulación
-    }
-
-    suspend fun updateUserPassword(userId: String, newPassword: String): Boolean {
-        // Lógica para actualizar el campo de contraseña (o alias privado) en Firestore
-        return true // Simulación
-    }
-
-    // Otros métodos para chats, hilos de foro, etc.
-}
