--- composeApp/src/commonMain/kotlin/org/connexuss/project/firebase/FirebaseFunctionsManager.kt
@@ -1,8 +1,0 @@
-package org.connexuss.project.firebase
-
-object FirebaseFunctionsManager {
-    suspend fun callFunction(name: String, data: Map<String, Any>): Any? {
-        // Lógica para invocar una función en Firebase Functions vía HTTP o SDK
-        return null // Simulación
-    }
-}
