--- composeApp/src/commonMain/kotlin/org/connexuss/project/firebase/FirebaseRemoteConfigHelper.kt
@@ -1,8 +1,0 @@
-package org.connexuss.project.firebase
-
-object FirebaseRemoteConfigHelper {
-    suspend fun fetchConfig(): Map<String, Any> {
-        // Lógica para obtener valores de Remote Config
-        return emptyMap()
-    }
-}
