+++ composeApp/src/commonMain/kotlin/org/connexuss/project/http/commonHttpClientFactory.kt
@@ -1,4 +1,4 @@
-package org.connexuss.project
+package org.connexuss.project.http
 /*
 import io.ktor.client.*
 import io.ktor.client.request.*
