+++ composeApp/src/commonMain/kotlin/org/connexuss/project/interfaces/Chats.kt
@@ -1,45 +1,6 @@
-package org.connexuss.project.interfaces.conversacion
+package org.connexuss.project.interfaces
 
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.Button
-import androidx.compose.material.Card
-import androidx.compose.material.Divider
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.OutlinedTextField
-import androidx.compose.material.Scaffold
-import androidx.compose.material.Text
-import androidx.compose.material.TopAppBar
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateListOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import kotlinx.datetime.Clock
-import kotlinx.datetime.TimeZone
-import kotlinx.datetime.toLocalDateTime
-import org.connexuss.project.comunicacion.Mensaje
-import org.connexuss.project.comunicacion.Conversacion
-import org.connexuss.project.interfaces.idiomas.traducir
-import org.connexuss.project.interfaces.sistema.DefaultTopBar
-import org.jetbrains.compose.ui.tooling.preview.Preview
-
+/*
 // --- Nuevo Chat ---
 @Composable
 @Preview
@@ -252,3 +213,4 @@ fun muestraChats() {
         }
     }
 }
+ */
