+++ composeApp/src/commonMain/kotlin/org/connexuss/project/interfaces/Foro.kt
@@ -1,4 +1,4 @@
-package org.connexuss.project.interfaces.foro
+package org.connexuss.project.interfaces
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -13,16 +13,10 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.material.Button
 import androidx.compose.material.Card
 import androidx.compose.material.Divider
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.OutlinedTextField
 import androidx.compose.material.Scaffold
 import androidx.compose.material.Text
-import androidx.compose.material.TopAppBar
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateListOf
@@ -33,9 +27,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
-import org.connexuss.project.interfaces.idiomas.traducir
-import org.connexuss.project.interfaces.modificadorTamannio.LimitaTamanioAncho
-import org.connexuss.project.interfaces.sistema.DefaultTopBar
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 // --- Foro ---
