+++ composeApp/src/commonMain/kotlin/org/connexuss/project/navegacion/Navegacion.kt
@@ -6,28 +6,28 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import org.connexuss.project.datos.UsuarioPrincipal
 import org.connexuss.project.datos.UsuariosPreCreados
-import org.connexuss.project.interfaces.ajustes.PantallaAjustesAyuda
-import org.connexuss.project.interfaces.ajustes.PantallaAjustesControlCuentas
-import org.connexuss.project.interfaces.colores.PantallaCambiarTema
-import org.connexuss.project.interfaces.colores.TemaConfig
-import org.connexuss.project.interfaces.foro.muestraForo
-import org.connexuss.project.interfaces.foro.muestraTemaForo
-import org.connexuss.project.interfaces.fuente.PantallaCambiarFuente
-import org.connexuss.project.interfaces.idiomas.PantallaIdiomas
-import org.connexuss.project.interfaces.sistema.PantallaEmailEnElSistema
-import org.connexuss.project.interfaces.sistema.PantallaEmailNoEnElSistema
-import org.connexuss.project.interfaces.sistema.PantallaLogin
-import org.connexuss.project.interfaces.sistema.PantallaRegistro
-import org.connexuss.project.interfaces.sistema.PantallaRestablecer
-import org.connexuss.project.interfaces.sistema.SplashScreen
-import org.connexuss.project.interfaces.sistema.mostrarChat
-import org.connexuss.project.interfaces.sistema.mostrarChatGrupo
-import org.connexuss.project.interfaces.sistema.mostrarPerfil
-import org.connexuss.project.interfaces.sistema.muestraAjustes
-import org.connexuss.project.interfaces.sistema.muestraChats
-import org.connexuss.project.interfaces.sistema.muestraContactos
-import org.connexuss.project.interfaces.sistema.muestraHomePage
-import org.connexuss.project.interfaces.sistema.muestraUsuarios
+import org.connexuss.project.interfaces.PantallaAjustesAyuda
+import org.connexuss.project.interfaces.PantallaAjustesControlCuentas
+import org.connexuss.project.interfaces.PantallaCambiarTema
+import org.connexuss.project.interfaces.TemaConfig
+import org.connexuss.project.interfaces.muestraForo
+import org.connexuss.project.interfaces.muestraTemaForo
+import org.connexuss.project.interfaces.PantallaCambiarFuente
+import org.connexuss.project.interfaces.PantallaIdiomas
+import org.connexuss.project.interfaces.PantallaEmailEnElSistema
+import org.connexuss.project.interfaces.PantallaEmailNoEnElSistema
+import org.connexuss.project.interfaces.PantallaLogin
+import org.connexuss.project.interfaces.PantallaRegistro
+import org.connexuss.project.interfaces.PantallaRestablecer
+import org.connexuss.project.interfaces.SplashScreen
+import org.connexuss.project.interfaces.mostrarChat
+import org.connexuss.project.interfaces.mostrarChatGrupo
+import org.connexuss.project.interfaces.mostrarPerfil
+import org.connexuss.project.interfaces.muestraAjustes
+import org.connexuss.project.interfaces.muestraChats
+import org.connexuss.project.interfaces.muestraContactos
+import org.connexuss.project.interfaces.muestraHomePage
+import org.connexuss.project.interfaces.muestraUsuarios
 
 @Composable
 fun Navegacion(
